move some primitives from NDLC.CLI to NDLC
