+++ NDLC/Constants.cs
@@ -1,6 +1,8 @@
-﻿namespace NDLC
+﻿using NBitcoin;
+
+namespace NDLC
 {
-	class Constants
+	public class Constants
 	{
 		public static readonly LockTime NeverLockTime = new LockTime(500000000 - 1);
 	}
